separating gpg_encryption and s3 service. moved encryption to a controller concern
