+++ app/config/application.rb
@@ -38,6 +38,7 @@ module IvCbvPayroll
     # Don't generate system test files.
     config.generators.system_tests = nil
     config.autoload_paths += %W[#{config.root}/lib]
+    config.autoload_paths += %W[#{config.root}/app/helpers]
 
     config.sites = SiteConfig.new(Rails.root.join("config", "site-config.yml"))
   end
