+++ app/spec/spec_helper.rb
@@ -1,12 +1,6 @@
 require 'webmock/rspec'
 require 'capybara/rspec'
 
-begin
-  module Cbv; end
-  require_relative '../app/controllers/concerns/cbv/payments_helper'
-rescue LoadError => e
-  puts "Warning: Unable to load Cbv::PaymentsHelper - #{e.message}"
-end
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -106,6 +100,4 @@ RSpec.configure do |config|
       @pinwheel ||= double('pinwheel')
     end
   end
-
-  config.include Cbv::PaymentsHelper if defined?(Cbv::PaymentsHelper)
 end
